Refactor provider utilities and switch validation caching to in-memory state
- move OpenAI-compatible response and embedding helpers out of `gemini.py` into new `llmproxy/openai.py`
- extract provider validation logging into new `llmproxy/utility.py` and update app startup to use it
- replace disk-based provider validation caching with process-lifetime in-memory caching and remove `LLM_PROXY_VALIDATION_CACHE` references from docs and gitignore
- convert `ProviderConfig` and `ProviderStatus` from dataclasses to frozen Pydantic models
- switch URL parsing and rewriting logic from `urllib.parse` to `yarl.URL` in validation, preflight, and provider ping scripts
- store trace `latency_ms` as rounded integers throughout the app and database model instead of floats
- add `yarl` dependency, bump `google-genai`, configure pytest warning filtering, and set Ruff line length to `130`
- add cache directories to `.gitignore` and remove checked-in `AGENTS.md`
- update tests and scripts to match the new models, integer latencies, URL handling, and formatting changes

+++ llmproxy/db.py
@@ -3,7 +3,7 @@ import re
 from datetime import datetime, timezone
 from os import getenv
 
-from sqlalchemy import DateTime, Float, Integer, String, Text, inspect, text
+from sqlalchemy import DateTime, Integer, String, Text, inspect, text
 from sqlalchemy.ext.asyncio import async_sessionmaker, create_async_engine
 from sqlalchemy.orm import DeclarativeBase, Mapped, mapped_column
 
@@ -27,19 +27,15 @@ class Trace(Base):
 
     id: Mapped[int] = mapped_column(Integer, primary_key=True, autoincrement=True)
     request_id: Mapped[str] = mapped_column(String(36), index=True)
-    created_at: Mapped[datetime] = mapped_column(
-        DateTime(timezone=True), default=lambda: datetime.now(timezone.utc)
-    )
+    created_at: Mapped[datetime] = mapped_column(DateTime(timezone=True), default=lambda: datetime.now(timezone.utc))
     provider: Mapped[str] = mapped_column(String(32))
     model: Mapped[str] = mapped_column(String(256))
     status_code: Mapped[int | None] = mapped_column(Integer, nullable=True)
-    latency_ms: Mapped[float | None] = mapped_column(Float, nullable=True)
+    latency_ms: Mapped[int | None] = mapped_column(Integer, nullable=True)
     request_json: Mapped[str | None] = mapped_column(Text, nullable=True)
     response_json: Mapped[str | None] = mapped_column(Text, nullable=True)
     error: Mapped[str | None] = mapped_column(Text, nullable=True)
-    idempotency_key: Mapped[str | None] = mapped_column(
-        String(256), nullable=True, index=True
-    )
+    idempotency_key: Mapped[str | None] = mapped_column(String(256), nullable=True, index=True)
     cache_key: Mapped[str | None] = mapped_column(String(64), nullable=True, index=True)
 
 
@@ -47,9 +43,7 @@ class CacheHit(Base):
     __tablename__ = "cache_hits"
 
     id: Mapped[int] = mapped_column(Integer, primary_key=True, autoincrement=True)
-    created_at: Mapped[datetime] = mapped_column(
-        DateTime(timezone=True), default=lambda: datetime.now(timezone.utc)
-    )
+    created_at: Mapped[datetime] = mapped_column(DateTime(timezone=True), default=lambda: datetime.now(timezone.utc))
     provider: Mapped[str] = mapped_column(String(32))
     model: Mapped[str] = mapped_column(String(256))
     cache_type: Mapped[str] = mapped_column(String(16))  # "response" or "idempotency"
@@ -73,28 +67,16 @@ async def init_db() -> None:
         trace_columns = await conn.run_sync(_get_trace_columns)
 
         if "idempotency_key" not in trace_columns:
-            await conn.execute(
-                text("ALTER TABLE traces ADD COLUMN idempotency_key VARCHAR(256)")
-            )
-        await conn.execute(
-            text(
-                "CREATE INDEX IF NOT EXISTS ix_traces_idempotency_key ON traces (idempotency_key)"
-            )
-        )
+            await conn.execute(text("ALTER TABLE traces ADD COLUMN idempotency_key VARCHAR(256)"))
+        await conn.execute(text("CREATE INDEX IF NOT EXISTS ix_traces_idempotency_key ON traces (idempotency_key)"))
         if "cache_key" not in trace_columns:
-            await conn.execute(
-                text("ALTER TABLE traces ADD COLUMN cache_key VARCHAR(64)")
-            )
-        await conn.execute(
-            text("CREATE INDEX IF NOT EXISTS ix_traces_cache_key ON traces (cache_key)")
-        )
+            await conn.execute(text("ALTER TABLE traces ADD COLUMN cache_key VARCHAR(64)"))
+        await conn.execute(text("CREATE INDEX IF NOT EXISTS ix_traces_cache_key ON traces (cache_key)"))
 
 
 async def test_db_connectivity() -> None:
     async with engine.connect() as conn:
         result = await conn.scalar(text("SELECT 1"))
     if result != 1:
-        raise RuntimeError(
-            f"Database connectivity check failed for {redact_db_url(DB_URL)!r}: expected 1, got {result!r}"
-        )
+        raise RuntimeError(f"Database connectivity check failed for {redact_db_url(DB_URL)!r}: expected 1, got {result!r}")
     LOGGER.info("Database connectivity check passed.")
